remove iothub_client_sample_mqtt from getstarted (#535)
* move f2s to Wstring.cpp, fix arduino do not support int64 which was used by json c, first fix of getstarted

* fix json-c for arduino do not support double

* .

* .

+++ AZ3166/src/cores/arduino/json-c/json_util.c
@@ -196,9 +196,9 @@ static void sscanf_is_broken_test()
 	}
 }
 
-int json_parse_int64(const char *buf, int64_t *retval)
+int json_parse_int64(const char *buf, int32_t *retval)
 {
-	int64_t num64;
+	int32_t num64;
 	const char *buf_sig_digits;
 	int orig_has_neg;
 	int saved_errno;
