+++ AZ3166/src/cores/arduino/json-c/json_util.h
@@ -73,7 +73,7 @@ extern int json_object_to_file_ext(const char *filename, struct json_object *obj
 const char *json_util_get_last_err(void);
 
 
-extern int json_parse_int64(const char *buf, int64_t *retval);
+extern int json_parse_int64(const char *buf, int32_t *retval);
 extern int json_parse_double(const char *buf, double *retval);
 
 /**
